# [2025/11/16] 修改BookViewModel的Price属性类型为decimal

+++ Models/ViewModels/BookViewModel.cs
@@ -33,7 +33,7 @@ namespace OnlineBookStore.Models.ViewModels
         public string CoverImageUrl { get; set; } = string.Empty;
 
         // 图书价格
-        public required float Price { get; set; }
+        public required decimal Price { get; set; }
 
         // 格式化后的图书价格字符串, 保留两位小数
         public string PriceString
